// lib/types.ts
export interface UserJwtPayload {
    userId: string;
}

export interface UserRecord {
    email: string;
    name: string;
    phone?: string;
    address?: string;
    usertype: number;
    shopid?: ShopBrief[];
    password?: string;
    tobeowner: boolean;
    created: string;
    tobeownerAt?: string;
}
// ...
export interface NoAuthorization{
    no_authorization: number;
}

export interface NoData{
    no_data: number;
}

export interface DuplicateEmail{
    duplicate_email: number;
}

export interface SetUserToOwnerDone{
    setusertoowner_done: number;
    id: string;
}

export interface SetShopOnboardDone{
    setshoponboard_done: number;
    id: string;
}

export interface ShopBlockUsers{
    shop_blockusers: number;
    shopId: string;
    userId: string;
}

export interface ShopUnBlockUsers{
    shop_unblockusers: number;
    shopId: string;
    userId: string;
}

export interface FileName{
    filename: string;
}

export interface ProcessResult{
    status: string;
    message: string;
}
// ...
export interface ShopBrief{
    id: string;
    shopname: string;
    onboard: boolean;
}